# server/practs/docs/mainRoute/pb/auth.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "AuthService",
    "description": "AuthService grpc, port: 50051.\n\nMainRoute: doc.json\nAuthService: http://localhost:3000/docs1/mainRoute/pb/auth.swagger.json\nProductService: http://localhost:3000/docs1/mainRoute/pb/product.swagger.json\nProfileService: http://localhost:3000/docs1/mainRoute/pb/profile.swagger.json",
    "version": "1.0"
  },
  "tags": [
    {
      "name": "AuthService"
    }
  ],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/authM/login": {
      "post": {
        "summary": "Login",
        "operationId": "AuthService_Login",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/authLoginResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/authLoginRequest"
            }
          }
        ],
        "tags": [
          "AuthService"
        ]
      }
    },
    "/authM/register": {
      "post": {
        "summary": "Registration",
        "operationId": "AuthService_Register",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/authRegisterResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/authRegisterRequest"
            }
          }
        ],
        "tags": [
          "AuthService"
        ]
      }
    },
    "/authM/validate": {
      "post": {
        "summary": "Validate",
        "operationId": "AuthService_Validate",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/authValidateResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/authValidateRequest"
            }
          }
        ],
        "tags": [
          "AuthService"
        ]
      }
    }
  },
  "definitions": {
    "authLoginRequest": {
      "type": "object",
      "properties": {
        "logname": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "authLoginResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "format": "int64"
        },
        "error": {
          "type": "string"
        },
        "token": {
          "type": "string"
        }
      }
    },
    "authRegisterRequest": {
      "type": "object",
      "properties": {
        "logname": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "authRegisterResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "format": "int64"
        },
        "error": {
          "type": "string"
        }
      }
    },
    "authValidateRequest": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        }
      }
    },
    "authValidateResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "format": "int64"
        },
        "error": {
          "type": "string"
        },
        "userId": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}